+++ back-end/src/models/postModels.ts
@@ -1,0 +1,12 @@
+export interface AddSensorProps {
+	name: string;
+	type: string;
+	unit: string;
+	readings?: [];
+}
+
+export interface ReadingDataProps {
+	sensorName: string;
+	time: number;
+	ML: number;
+}
